requirements.txt
mysql-connector-python
openai
pydantic
pydantic-settings
langchain
langchain-openai
fastapi[standard]